+++ lib/src/undo/history_manager.dart
@@ -1,6 +1,6 @@
 import '../action/base_action.dart';
 import '../data_store.dart';
-import 'action_event.dart';
+import '../action/action_event.dart';
 
 class KiwiHistoryManager {
   final KiwiWatermelonDataStore store;
